Use fog-aws instead of the base fog gem.
Fog has been splitting out it's providers into separate gems and the
`fog` gem is becoming a metapackage for all of the providers. Since
pgbackups-archive is hard coded to use AWS it should use the (recently
released) fog-aws gem instead.

This has the advantage of only needing to pull in those pieces of fog
that are actually used and not having to pull in 10-20 other gems that
fog depends on.